* intel-fix-msvc:
  examples/build-msvc: re-add dropped Makefile
  examples/build-unix: re-add dropped Makefile
  Fix MSBUILD-based batch example execution
  TLM examples: Update MSVC command-line build setup
  vsvars.bat: sync with SystemC regressions repository
  Cleanup MSVC project files for SystemC/TLM examples
  Remove MSVC warnings in (or triggered from) SystemC/TLM examples
  Update GNU make setup for SystemC & TLM examples
  Automake: update examples to include updated MSVC setup
  sc_vector: add missing SC_API annotation
  MSVC: Avoid warning C4267 (model side)
  VCD/WIF: hide MSVC warning
  sc_signal: cleanups for MSVC (DLL) build

+++ src/sysc/communication/sc_signal_ports.h
@@ -1211,7 +1211,7 @@ void
 sc_inout<T>::remove_traces() const
 {
     if( m_traces != 0 ) {
-		for( int i = m_traces->size() - 1; i >= 0; -- i ) {
+		for( int i = static_cast<int>(m_traces->size()) - 1; i >= 0; -- i ) {
 	        delete (*m_traces)[i];
 		}
 		delete m_traces;
